Adds props, turns rest of app TypeScript
