+++ src/containers/Signup/Component.tsx
@@ -4,13 +4,19 @@ import Typography from '@material-ui/core/Typography';
 import Grid from '@material-ui/core/Grid';
 import Link from '@material-ui/core/Link';
 import InputAdornment from '@material-ui/core/InputAdornment';
+import IconButton from '@material-ui/core/IconButton';
+import Visibility from '@material-ui/icons/VisibilityOutlined';
+import VisibilityOff from '@material-ui/icons/VisibilityOffOutlined';
 import { makeStyles } from '@material-ui/core/styles';
 import { Link as RouterLink, Redirect } from 'react-router-dom';
 import EmailIcon from '@material-ui/icons/EmailOutlined';
-import InputBase from '@material-ui/core/InputBase';
-import * as constants from '../../constants.ts';
+import PasswordIcon from '@material-ui/icons/VpnKeyOutlined';
 
-import AppWindow from '../../components/AppWindow/index.ts';
+import InputBase from '@material-ui/core/InputBase';
+
+import * as constants from '../../constants';
+import AppWindow from '../../components/AppWindow/index';
+import { DispatchProps, StateProps } from './Container';
 
 const useStyles = makeStyles((theme) => ({
   header: {
@@ -26,7 +32,21 @@ const useStyles = makeStyles((theme) => ({
     height: '48px',
     overflow: 'hidden',
     borderWidth: '0.5px',
+    borderBottomWidth: '0.5px',
+    borderBottomStyle: 'dashed',
     borderStyle: 'solid',
+    borderColor: '#C0C0C0',
+  },
+  password: {
+    marginBottom: theme.spacing(3),
+    background: 'white',
+    borderBottomRightRadius: '4px',
+    borderBottomLeftRadius: '4px',
+    borderWidth: '0.5px',
+    borderTopWidth: '0px',
+    borderStyle: 'solid',
+    height: '48px',
+    overflow: 'hidden',
     borderColor: '#C0C0C0',
   },
   continueButton: {
@@ -35,14 +55,15 @@ const useStyles = makeStyles((theme) => ({
   },
   loginLink: {
     marginBottom: theme.spacing(3),
+    cursor: 'pointer',
   },
   content: {
     width: '100%',
   },
   errorMsg: {
     marginBottom: theme.spacing(3),
-    paddingLeft: theme.spacing(3),
-    paddingRight: theme.spacing(3),
+    paddingLeft: '24px',
+    paddingRight: '24px',
   },
   inputRoot: {
     backgroundColor: 'white',
@@ -63,34 +84,31 @@ const useStyles = makeStyles((theme) => ({
 
 
 const Login = ({
-  isAuthorized, sendResetPasswordEmailRequestState, sendResetPasswordEmail, resetUser,
-  error,
-}) => {
+  isLoading, isError, isSuccess, signupWithEmail,
+  googleSignIn, isAuthorized, error, resetUser,
+}: StateProps & DispatchProps) => {
 
+  const [password, setPassword] = useState('');
+  const [email, setEmail] = useState('');
+  const [showPassword, setShowPassword] = useState(false);
 
   React.useEffect(() => resetUser, []);
 
-  const [email, setEmail] = useState('');
+  const isDisabled = isLoading || isSuccess;
 
-  const isDisabled = (
-    sendResetPasswordEmailRequestState === constants.LOADING ||
-    sendResetPasswordEmailRequestState === constants.SUCCESS
-  );
 
-  const submitDisabled = isDisabled || !email;
+  const submitDisabled = isDisabled || !password || password.length < 7 || !email;
 
   const classes = useStyles();
 
   return (
     <AppWindow size="xs">
-
       {
         isAuthorized &&
         <Redirect
           to='/'
         />
       }
-
       <Grid className={classes.content} alignItems="center" direction="column" container>
         <Grid className={classes.header} item>
           <Typography align="center" variant="h4">{constants.APP_NAME}</Typography>
@@ -98,17 +116,46 @@ const Login = ({
         <Grid item>
           <InputBase
             name="email"
+            autoComplete="username"
             style={{ display: 'flex' }}
             className={classes.email}
             onChange={(e) => setEmail(e.target.value)}
-            // label="Type your email"
             disabled={isDisabled}
             placeholder="email"
             value={email}
-            variant="filled"
             startAdornment={
               <InputAdornment position="start">
                 <EmailIcon className={classes.inputStartAdornment} />
+              </InputAdornment>
+            }
+          />
+          <InputBase
+            autoComplete="new-password"
+            name="password"
+            className={classes.password}
+            onChange={(e) => setPassword(e.target.value)}
+            disabled={isDisabled}
+            type={showPassword ? 'text' : 'password'}
+            value={password}
+            placeholder="password"
+            startAdornment={
+              <InputAdornment position="start">
+                <PasswordIcon className={classes.inputStartAdornment} />
+              </InputAdornment>
+            }
+            endAdornment={
+              <InputAdornment position="end">
+                <IconButton
+                  aria-label="toggle password visibility"
+                  onClick={() => setShowPassword(!showPassword)}
+                  onMouseDown={() => setShowPassword(!showPassword)}
+                >
+                  {
+                    showPassword ?
+                      (<Visibility className={classes.inputAdornment} />) :
+                      (<VisibilityOff className={classes.inputAdornment} />)
+                  }
+                </IconButton>
               </InputAdornment>
             }
           />
@@ -117,14 +164,14 @@ const Login = ({
           <Button
             className={classes.continueButton}
             disabled={submitDisabled}
-            onClick={() => sendResetPasswordEmail(email)}
+            onClick={() => signupWithEmail(email, password)}
             variant="contained"
             color="primary"
-          >Send password reset email</Button>
+          >Create account</Button>
         </Grid>
 
         {
-          sendResetPasswordEmailRequestState === constants.ERROR && (
+          isError && (
             <Grid className={classes.errorMsg} item>
               <Typography color="error" variant="body1">
                 {(error && error.message) || 'Something went wrong!'}
@@ -132,6 +179,20 @@ const Login = ({
             </Grid>
           )
         }
+
+        <Grid
+          className={classes.loginLink}
+          item
+        >
+          <Link
+            onClick={googleSignIn}
+          >
+            <Typography variant="body2">
+              Continue with Google
+            </Typography>
+          </Link>
+        </Grid>
+
         <Grid
           className={classes.loginLink}
           item
@@ -141,7 +202,7 @@ const Login = ({
             to={'/'}
           >
             <Typography variant="body2">
-              Go back to login
+              Already have an account? Login instead.
             </Typography>
           </Link>
         </Grid>
