+++ src/containers/__tests__/Signup.spec.tsx
@@ -5,22 +5,27 @@ import { HashRouter as Router } from 'react-router-dom';
 // eslint-disable-next-line no-unused-vars
 import { createMuiTheme, MuiThemeProvider } from '@material-ui/core/styles';
 
-import Component from '../Warband/Component.tsx';
+import Component from '../Signup/Component';
 
-
-jest.mock('@material-ui/lab/Autocomplete', () => () => 'autocomplete');
 
 const theme = createMuiTheme({});
 
 
-describe('Warband', () => {
+describe('Signup', () => {
   it('Should render (smoke test)', () => {
     const { asFragment } = render(
+      // Requires to be wrapped in router because NotFoundPage uses react-router-doms Link element
+      // Requires to be wrapped MuiThemeProvider because NotFoundPage is using Theme from material-ui
       <Router>
         <MuiThemeProvider theme={theme}>
           <Component
-            addWarbandReset={() => { }}
-            getWarbands={() => { }}
+            resetUser={() => { }}
+            signupWithEmail={() => { }}
+            googleSignIn={() => { }}
+            isLoading={false}
+            isError={false}
+            isSuccess={false}
+            isAuthorized={false}
           />
         </MuiThemeProvider>
       </Router>
