+++ src/containers/__tests__/Warband.spec.tsx
@@ -5,22 +5,26 @@ import { HashRouter as Router } from 'react-router-dom';
 // eslint-disable-next-line no-unused-vars
 import { createMuiTheme, MuiThemeProvider } from '@material-ui/core/styles';
 
-import Component from '../ForgotPassword/Component';
+import Component from '../Warband/Component';
+
+
+jest.mock('@material-ui/lab/Autocomplete', () => () => 'autocomplete');
 
 const theme = createMuiTheme({});
 
 
-describe('Forgot Password', () => {
+describe('Warband', () => {
   it('Should render (smoke test)', () => {
     const { asFragment } = render(
-      // Requires to be wrapped in router because NotFoundPage uses react-router-doms Link element
-      // Requires to be wrapped MuiThemeProvider because NotFoundPage is using Theme from material-ui
       <Router>
         <MuiThemeProvider theme={theme}>
           <Component
-            isAuthorized={false}
-            sendResetPasswordEmailRequestState=""
-            sendResetPasswordEmail={() => {}}
+            isSuccessGetWarbands={false}
+            logout={() => { }}
+            saveWarband={() => { }}
+            removeWarband={() => { }}
+            addWarbandReset={() => { }}
+            getWarbands={() => { }}
           />
         </MuiThemeProvider>
       </Router>
